pyproject.toml
[project]
name = "wonttrade"
version = "0.1.0"
description = "Autonomous LLM-driven trading daemon for Hyperliquid"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "hyperliquid-python-sdk>=0.20.0",
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
